Tolerate extra VolumeMounts in Pod Containers. (#23)
When managing Pods directly, as opposed to filling in a Pod Template in
something like a Deployment, we need to tolerate extra VolumeMounts
being added to our Containers. The API server adds a mount to every
Container for the ServiceAccount token, for example.

At the same time, we need to notice when one of the "extra" VolumeMounts
is actually one that we put there ourselves previously, since we need to
remove it instead of leaving it untouched.

This change extends a pattern that we borrowed from Deployment itself to
handle our Annotations to also apply to VolumeMounts in Containers.

The trick is to set an Annotation whose value depends on the unordered
set of Containers/VolumeMounts we ourselves care about. If that set ever
changes, the Annotation changes, and our Pod will have to be deleted and
recreated, effectively removing anything we no longer want.

Instead of storing the full list of Containers and VolumeMounts that we
previously wanted (to distinguish them from ones that someone else
added), we simply store a hash of these lists, since all we care about
is comparing for equality.

Signed-off-by: Anthony Yeh <enisoc@planetscale.com>