deploy, pkg: mount mycnf and vtbackup/vttablet initialization script from configmaps
Currently vtbackup and vttablet pods run a script from an init container
which copies binaries and certs, and creates some extra mycnf files. The
contents of this script and these mycnfs are hardcoded in Golang code.

This commit is a refactor which moves the contents of these mycnfs and
initialization script into K8s configmaps. The initialization script is
shared by vtbackup and vttablet, and parameterized with environment
variables.

Deploy notes:

 * Updating the operator image will cause vttablet pods to be recreated.
 * The configmaps must be deployed before updating the operator image,
   otherwise vtbackup/vttablet pods won't be able to mount them.

Signed-off-by: Max Englander <max@planetscale.com>