Set vttablet init container resource requests and limits (#370)
* Set vttablet init container resource requests and limits

Sets the resource requirements on each of the default vttablet init containers to the same request and limit values as the vttablet container itself.

This fixes the issue reported in #212 where pods would fail to be created on clusters with strict quota requirements because no limit was set on the init containers.

## History

This is a reimplementation of #216 with the change that it no longer sets a default resource _limit_.

This fixes the issue reported in #240 and discussed in #244 where the vttablet pod would fail to be created when the user sets a request larger than the default limit without also setting a larger limit. The user's nil limit would not override the default value because `update.ResourceRequirements(dst, src)` won't update `dst` with a nil `src` ([for good reasons](https://github.com/planetscale/vitess-operator/blob/da7efd4c945193864f5996cb793055449c4de97a/pkg/operator/update/podspec.go#L201-L203)).

With this change, users can once again increase the request by itself. If a cluster requires a limit, the user will already need to have set it on for vttablet so we can use the same value to override the init containers' default nil limit.

fixes #212
fixes #240

Signed-off-by: J. Brandt Buckley <brandt@runlevel1.com>

* refactor: move defaults to the correct location

Signed-off-by: Manan Gupta <manan@planetscale.com>

---------

Signed-off-by: J. Brandt Buckley <brandt@runlevel1.com>
Signed-off-by: Manan Gupta <manan@planetscale.com>
Co-authored-by: J. Brandt Buckley <brandt@runlevel1.com>