Use cache mount in container image builds
Without a cache mount, any change to the repository results in a new
image build, even if there are no changes to the operator's codebase.
Not only each build takes over 40 seconds (on my machine), it also
creates about 4 GB of Docker layers, every time. This adds up quickly
and can fill up the Docker disk allocation after only a handful of runs
on a Mac.

This change drops the image build time from 40+ to 2 seconds. The layers
that store the Go module and build caches are reused across builds.
As a result, the disk growth between builds is not nearly as noticable.

It should be a completely transparent change locally and in CI.

Signed-off-by: Sergey Stankevich <stankevich@users.noreply.github.com>