Simplify backup storage and permissions
Using `hostPath` as a volume in Kind adds a bit of complexity. A configuration file must be generated, a directory must be pre-created, permissions need to be set, a cleanup step is required when the cluster is destroyed, etc. This isn't as big of a deal as correcting permissions on backup subdirectories midway through test runs only to allow the backup subcontroller "see" the backups and surface them in the `kubectl get vitessbackups` output. This is pretty fragile and is prone to edge cases.

A Kubernetes-native way to deal with this issue is setting `securityContext.fsGroup` on all workloads involved in taking and reading backups. Unfortunately, `fsGroup` doesn't work with `hostPath` in Kubernetes or Kind (https://github.com/rancher/local-path-provisioner/issues/41#issuecomment-907392513, https://github.com/kubernetes-sigs/kind/issues/830). Additionally, using `hostPath` volumes is generally discouraged (https://kubernetes.io/docs/concepts/storage/volumes/#hostpath).

Hence this commit. It replaces `hostPath` volumes with `persistentVolumeClaim` volumes. This makes `fsGroup` work and eliminates all `chmod`-ing. It also removes Kind config generation and host-container directory sharing. The teardown process also becomes simpler, as backup files are automatically removed during Kind cluster deletion. Lastly, this setup is slightly more representative of real-world usage.

To make it work, `securityContext.fsGroup` of the operator Deployment must be set to the `fsGroup` value of Vitess workloads, which is 999 by default. The backup subcontroller then inherits the `fsGroup` of the operator. Otherwise, it runs as UID 1001, which doesn't have permissions to read Vitess backup files from the mounted persistent volume.

Signed-off-by: Sergey Stankevich <stankevich@users.noreply.github.com>